+++ run-tutorial.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-echo "Preparing and starting the Fulcro Tutorial..."
-echo "As soon as you see the compilation complete you can navigate to http://localhost:8080/tutorial.html"
-./run-figwheel.sh tutorial