+++ src/book/book/forms/forms_demo_3.cljs
@@ -4,8 +4,7 @@
     [fulcro.client.dom :as dom]
     [fulcro.client.primitives :as prim :refer [defsc]]
     [fulcro.client.mutations :as m :refer [defmutation]]
-    [fulcro.ui.forms :as f :refer [defvalidator]]
-    [fulcro.i18n :refer [tr]]))
+    [fulcro.ui.forms :as f :refer [defvalidator]]))
 
 (declare ValidatedPhoneForm)