+++ src/main/deps.cljs
@@ -1,5 +1,0 @@
-{:foreign-libs
-          [{:file     "yahoo/intl-messageformat-with-locales.js"
-            :file-min "yahoo/intl-messageformat-with-locales.min.js"
-            :provides ["yahoo.intl-messageformat-with-locales"]}]}
-