+++ src/main/fulcro/client/dom.cljs
@@ -3,6 +3,7 @@
   (:require-macros [fulcro.client.dom :as dom])
   (:require [cljsjs.react]
             [cljsjs.react.dom]
+            fulcro.util
             [goog.object :as gobj]))
 
 (defn- update-state